fix: enhance IDE autocompletion
This commit introduces explicit delegation to improve convenience and support for IDE autocompletion. The following delegates have been added:

- `browser` object delegate
- `page` object delegate
- `report(String)` method delegate
- `getPageSource()` method delegate
- Refactor `ContainerAwareDownloadSupport` to a more generic support trait and move `createFileInputSource` method to it.
- Move support classes to `support` package.
